Move modals to routes folder
