+++ client/src/common/components/ProfileInfo/ProfileInfo.tsx
@@ -1,12 +1,12 @@
 import React, {useState} from 'react';
 import {useTranslation} from 'react-i18next';
 import styled from 'styled-components/native';
-import Button from '../../../common/components/Buttons/Button';
-import {Spacer16} from '../../../common/components/Spacers/Spacer';
-import Input from '../../../common/components/Typography/TextInput/TextInput';
-import {SPACINGS} from '../../../common/constants/spacings';
-import useChangeProfileInfo from '../../Profile/hooks/useChangeProfileInfo';
-import ProfilePicture from '../../../common/components/User/ProfilePicture';
+import Button from '../Buttons/Button';
+import {Spacer16} from '../Spacers/Spacer';
+import Input from '../Typography/TextInput/TextInput';
+import {SPACINGS} from '../../constants/spacings';
+import useChangeProfileInfo from '../../../routes/Profile/hooks/useChangeProfileInfo';
+import ProfilePicture from '../User/ProfilePicture';
 import useUser from '../../../lib/user/hooks/useUser';
 
 const Container = styled.View({
