+++ client/src/lib/navigation/Stacks.tsx
@@ -13,13 +13,13 @@ import ChangingRoom from '../../routes/Session/ChangingRoom';
 import IntroPortal from '../../routes/Session/IntroPortal';
 import OutroPortal from '../../routes/Session/OutroPortal';
 import DailyProvider from '../daily/DailyProvider';
-import SessionModal from '../../routes/Sessions/components/SessionModal';
-import CreateSessionModal from '../../routes/Sessions/components/CreateSessionModal';
-import JoinSessionModal from '../../routes/Sessions/components/JoinSessionModal';
+import SessionModal from '../../routes/SessionModal/SessionModal';
+import CreateSessionModal from '../../routes/CreateSessionModal/CreateSessionModal';
+import JoinSessionModal from '../../routes/JoinSessionModal/JoinSessionModal';
 import useNavigationState from './state/state';
-import UpgradeAccountModal from '../../routes/Profile/UpgradeAccountModal';
-import AddSessionModal from '../../routes/Sessions/components/AddSessionModal';
-import SessionUnavailableModal from '../../routes/Sessions/components/SessionUnavailableModal';
+import UpgradeAccountModal from '../../routes/UpgradeAccountModal/UpgradeAccountModal';
+import AddSessionModal from '../../routes/AddSessionModal/AddSessionModal';
+import SessionUnavailableModal from '../../routes/SessionUnavailableModal/SessionUnavailableModal';
 
 const RootStack = createNativeStackNavigator<RootStackProps>();
 const SessionStack = createNativeStackNavigator<SessionStackProps>();
