+++ client/src/routes/AddSessionModal/AddSessionModal.tsx
@@ -3,17 +3,17 @@ import {useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {useTranslation} from 'react-i18next';
 
-import Gutters from '../../../common/components/Gutters/Gutters';
-import HalfModal from '../../../common/components/Modals/HalfModal';
+import Gutters from '../../common/components/Gutters/Gutters';
+import HalfModal from '../../common/components/Modals/HalfModal';
 import {
   Spacer16,
   Spacer24,
   Spacer8,
-} from '../../../common/components/Spacers/Spacer';
-import {Display24} from '../../../common/components/Typography/Display/Display';
+} from '../../common/components/Spacers/Spacer';
+import {Display24} from '../../common/components/Typography/Display/Display';
 
-import {ModalStackProps} from '../../../lib/navigation/constants/routes';
-import Button from '../../../common/components/Buttons/Button';
+import {ModalStackProps} from '../../lib/navigation/constants/routes';
+import Button from '../../common/components/Buttons/Button';
 
 const AddSessionModal: React.FC = () => {
   const {t} = useTranslation('Component.AddSessionModal');
