+++ client/src/routes/JoinSessionModal/JoinSessionModal.tsx
@@ -4,20 +4,20 @@ import React, {useState} from 'react';
 import {useTranslation} from 'react-i18next';
 import styled from 'styled-components/native';
 
-import {JoinSessionError} from '../../../../../shared/src/errors/Session';
-import {COLORS} from '../../../../../shared/src/constants/colors';
+import {JoinSessionError} from '../../../../shared/src/errors/Session';
+import {COLORS} from '../../../../shared/src/constants/colors';
 
-import Gutters from '../../../common/components/Gutters/Gutters';
-import HalfModal from '../../../common/components/Modals/HalfModal';
-import {Spacer16, Spacer24} from '../../../common/components/Spacers/Spacer';
-import {Body14} from '../../../common/components/Typography/Body/Body';
-import {Display24} from '../../../common/components/Typography/Display/Display';
-import VerificationCode from '../../Profile/components/VerificationCode';
+import Gutters from '../../common/components/Gutters/Gutters';
+import HalfModal from '../../common/components/Modals/HalfModal';
+import {Spacer16, Spacer24} from '../../common/components/Spacers/Spacer';
+import {Body14} from '../../common/components/Typography/Body/Body';
+import {Display24} from '../../common/components/Typography/Display/Display';
+import VerificationCode from '../../common/components/VerificationCode/VerificationCode';
 
-import {ModalStackProps} from '../../../lib/navigation/constants/routes';
+import {ModalStackProps} from '../../lib/navigation/constants/routes';
 
-import {joinSession} from '../api/session';
-import useSessions from '../hooks/useSessions';
+import {joinSession} from '../Sessions/api/session';
+import useSessions from '../Sessions/hooks/useSessions';
 
 const ErrorText = styled(Body14)({color: COLORS.ERROR, textAlign: 'center'});
