+++ client/src/routes/UpgradeAccountModal/UpgradeAccountModal.tsx
@@ -9,11 +9,11 @@ import Input from '../../common/components/Typography/TextInput/TextInput';
 import {Body14, Body16} from '../../common/components/Typography/Body/Body';
 import {Heading16} from '../../common/components/Typography/Heading/Heading';
 import Gutters from '../../common/components/Gutters/Gutters';
-import {requestPromotion, verifyPromotion} from './api/user';
+import {requestPromotion, verifyPromotion} from '../Profile/api/user';
 import styled from 'styled-components/native';
 import HalfModal from '../../common/components/Modals/HalfModal';
 import useUserState from '../../lib/user/state/state';
-import VerificationCode from './components/VerificationCode';
+import VerificationCode from '../../common/components/VerificationCode/VerificationCode';
 import {ModalStackProps} from '../../lib/navigation/constants/routes';
 import {VerificationError} from '../../../../shared/src/errors/User';
 import {COLORS} from '../../../../shared/src/constants/colors';
